# [wip] perf(ext/ffi): Low overhead v8::Value <-> FFI (de)serialization