--- a/ext/ffi/00_ffi.js
+++ b/ext/ffi/00_ffi.js
@@ -134,7 +134,9 @@
     }
 
     static of(typedArray) {
-      return new UnsafePointer(core.opSync("op_ffi_ptr_of", typedArray));
+      return new UnsafePointer(
+        unpackU64(core.opSync("op_ffi_ptr_of", typedArray)),
+      );
     }
 
     valueOf() {
@@ -144,37 +146,11 @@
   const UnsafePointerPrototype = UnsafePointer.prototype;
 
   function prepareArgs(types, args) {
-    const parameters = [];
-    const buffers = [];
-
-    for (let i = 0; i < types.length; i++) {
-      const type = types[i];
-      const arg = args[i];
-
-      if (type === "pointer") {
-        if (
-          ObjectPrototypeIsPrototypeOf(ArrayBufferPrototype, arg?.buffer) &&
-          arg.byteLength !== undefined
-        ) {
-          parameters.push(buffers.length);
-          buffers.push(arg);
-        } else if (ObjectPrototypeIsPrototypeOf(UnsafePointerPrototype, arg)) {
-          parameters.push(packU64(arg.value));
-          buffers.push(undefined);
-        } else if (arg === null) {
-          parameters.push(null);
-          buffers.push(undefined);
-        } else {
-          throw new TypeError(
-            "Invalid ffi arg value, expected TypedArray, UnsafePointer or null",
-          );
-        }
-      } else {
-        parameters.push(arg);
-      }
-    }
-
-    return { parameters, buffers };
+    return args.map((arg) =>
+      ObjectPrototypeIsPrototypeOf(UnsafePointerPrototype, arg)
+        ? packU64(arg.value)
+        : arg
+    );
   }
 
   class UnsafeFnPointer {
@@ -187,33 +163,31 @@
     }
 
     call(...args) {
-      const { parameters, buffers } = prepareArgs(
+      const parameters = prepareArgs(
         this.definition.parameters,
         args,
       );
       if (this.definition.nonblocking) {
-        const promise = core.opAsync("op_ffi_call_ptr_nonblocking", {
-          pointer: packU64(this.pointer.value),
-          def: this.definition,
+        const promise = core.opAsync(
+          "op_ffi_call_ptr_nonblocking",
+          packU64(this.pointer.value),
+          this.definition,
           parameters,
-          buffers,
-        });
-
+        );
         if (this.definition.result === "pointer") {
-          return promise.then((value) => new UnsafePointer(value));
+          return promise.then((value) => new UnsafePointer(unpackU64(value)));
         }
 
         return promise;
       } else {
-        const result = core.opSync("op_ffi_call_ptr", {
-          pointer: packU64(this.pointer.value),
-          def: this.definition,
+        const result = core.opSync(
+          "op_ffi_call_ptr",
+          packU64(this.pointer.value),
+          this.definition,
           parameters,
-          buffers,
-        });
-
+        );
         if (this.definition.result === "pointer") {
-          return new UnsafePointer(result);
+          return new UnsafePointer(unpackU64(result));
         }
 
         return result;
@@ -246,10 +220,11 @@
               type,
             },
           );
-          if (type === "pointer") {
-            value = new UnsafePointer(value);
-          } else if (type === "u64") {
+          if (type === "pointer" || type === "u64") {
             value = unpackU64(value);
+            if (type === "pointer") {
+              value = new UnsafePointer(value);
+            }
           } else if (type === "i64") {
             value = unpackI64(value);
           }
@@ -269,36 +244,30 @@
         const types = symbols[symbol].parameters;
 
         const fn = (...args) => {
-          return core.opSync("op_ffi_call", this.#rid, symbol, args);
-          const { parameters, buffers } = prepareArgs(types, args);
-
+          const parameters = prepareArgs(types, args);
           if (isNonBlocking) {
-            const promise = core.opAsync("op_ffi_call_nonblocking", {
-              rid: this.#rid,
-              symbol,
+            const promise = core.opAsync(
+              "op_ffi_call_nonblocking",
+              this.#rid,
+              symbols,
               parameters,
-              buffers,
-            });
-
+            );
             if (symbols[symbol].result === "pointer") {
               return promise.then((value) =>
-                new UnsafePointer(value),
+                new UnsafePointer(unpackU64(value))
               );
             }
-
             return promise;
           } else {
-            const result = core.opSync("op_ffi_call", {
-              rid: this.#rid,
+            const result = core.opSync(
+              "op_ffi_call",
+              this.#rid,
               symbol,
               parameters,
-              buffers,
-            });
-
+            );
             if (symbols[symbol].result === "pointer") {
-              return new UnsafePointer(result);
+              return new UnsafePointer(unpackU64(result));
             }
-
             return result;
           }
         };