--- a/test_ffi/src/lib.rs
+++ b/test_ffi/src/lib.rs
@@ -114,6 +114,9 @@ pub extern "C" fn get_sleep_blocking_ptr() -> *const c_void {
 }
 
 #[no_mangle]
+pub extern "C" fn nop() {}
+
+#[no_mangle]
 pub static static_u32: u32 = 42;
 
 #[repr(C)]