Fixes https://github.com/denoland/deno/issues/14431

TODO: 
- [x] Get's rid of `serde_json::Value`.
- [x] FFI calls are no longer 'ops' (not auto-generated with deno_ops)
- [ ] Throw errors in JS (don't `unwrap` lol)
- [ ] `tokio::spawn_blocking` for async calls.

Benchmarks -

Deno 1.12.1:
```
cpu: Apple M1
runtime: deno 1.21.1 (aarch64-apple-darwin)

file:///Users/divy/gh/deno/test_ffi/bench.js
benchmark      time (avg)             (min … max)       p75       p99      p995
------------------------------------------------- -----------------------------
nop()      456.78 ns/iter (440.37 ns … 903.48 ns) 445.07 ns 727.85 ns 903.48 ns
add(1,2)   678.34 ns/iter (662.85 ns … 694.42 ns) 681.14 ns 694.42 ns 694.42 ns
```

This PR:
```
cpu: Apple M1
runtime: deno 1.21.2 (aarch64-apple-darwin)

file:///Users/divy/gh/deno/test_ffi/bench.js
benchmark      time (avg)             (min … max)       p75       p99      p995
------------------------------------------------- -----------------------------
nop()      152.12 ns/iter  (144.67 ns … 261.4 ns) 150.43 ns 232.65 ns  252.9 ns
add(1,2)   260.75 ns/iter (253.98 ns … 268.91 ns) 261.96 ns 267.75 ns 267.81 ns
```