@littledivy 2022-05-10
> Why can this not be done with ops + [`serde_v8::Value`](https://docs.rs/serde_v8/latest/serde_v8/struct.Value.html)?

@lucacasonato We need the handle scope too.